Moves the RTE related classes to the rte package.
